Changes CSS var colors, changed default card
Changed the appearance of the default card -- by default, has a lighter
color scheme, when the book has been read, the color scheme will invert
and get darker. I also added a removable and addable class that can be
toggled later on using CSS. The primary, secondary, and tertiary colors
for the webpage were modified as well.